Whistle tuners now work, but still open the locomotive menu, while RC's tuner doesn't.

+++ src/main/java/net/mrhitech/firmarail/util/mixininterface/LocomotiveInterface.java
@@ -1,6 +1,6 @@
 package net.mrhitech.firmarail.util.mixininterface;
 
 public interface LocomotiveInterface {
-    int getNewWhistlePitch();
+    float getNewWhistlePitch();
     void whistle();
 }
